+++ core/core/Cargo.toml
@@ -133,7 +133,6 @@ services-sftp = ["dep:openssh", "dep:openssh-sftp-client", "dep:fastpool"]
 services-sqlite = ["dep:sqlx", "sqlx?/sqlite", "dep:ouroboros"]
 services-surrealdb = ["dep:surrealdb"]
 services-swift = []
-services-tikv = ["tikv-client"]
 services-upyun = ["dep:hmac", "dep:sha1"]
 services-vercel-artifacts = []
 services-webdav = []
@@ -237,8 +236,6 @@ redis = { version = "1.0", features = [
 ], optional = true }
 # for services-rocksdb
 rocksdb = { version = "0.21.0", default-features = false, optional = true }
-# for services-tikv
-tikv-client = { version = "0.3.0", optional = true, default-features = false }
 # for services-hdfs-native
 hdfs-native = { version = "0.13", optional = true }
 # for services-surrealdb
