+++ core/src/lib.rs
@@ -57,6 +57,8 @@ pub mod services {
     pub use opendal_service_s3::*;
     #[cfg(feature = "services-sled")]
     pub use opendal_service_sled::*;
+    #[cfg(feature = "services-tikv")]
+    pub use opendal_service_tikv::*;
     #[cfg(feature = "services-vercel-blob")]
     pub use opendal_service_vercel_blob::*;
 }
